+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- method to delete a file from storage with `DeleteFile`.
+
 ### Fixed
 - support [MO2](https://github.com/ModOrganizer2/modorganizer/).